+++ CHANGES.md
@@ -1,5 +1,6 @@
 # 0.1.0
 * First version
+* Support for interacting with checks
 * Support for interacting with clients
 * Support for interacting with events
 * Support for interacting with stashes